requirements/common.txt
Django==1.7.4
defusedxml==0.4.1
dj-static==0.0.6
django-allauth==0.19.1
django-bootstrap-form==3.1
django-custom-user==0.5
django-enumfield==1.2
django-tastypie==0.12.1
mimeparse==0.1.3
oauthlib==0.7.2
psycopg2==2.5.4
python-dateutil==2.4.0
python-mimeparse==0.1.4
python3-openid==3.0.5
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
static3==0.5.1
